+++ tests/TestCase.php
@@ -1,38 +1,11 @@
-<?php declare(strict_types=1);
+<?php
 
-/**
- * Copyright (C) Brian Faust
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
+declare(strict_types=1);
 
 namespace Tests;
 
-use Illuminate\Foundation\Bootstrap\LoadEnvironmentVariables;
-use Orchestra\Testbench\TestCase as BaseTestCase;
-use Override;
-use Spatie\LaravelData\LaravelDataServiceProvider;
+use PHPUnit\Framework\TestCase as BaseTestCase;
 
-/**
- * @internal
- *
- * @author Brian Faust <brian@cline.sh>
- */
 abstract class TestCase extends BaseTestCase
 {
-    protected function defineEnvironment($app): void
-    {
-        $app->useEnvironmentPath(__DIR__.'/..');
-
-        $app->bootstrapWith([LoadEnvironmentVariables::class]);
-    }
-
-    #[Override()]
-    protected function getPackageProviders($app)
-    {
-        return [
-            LaravelDataServiceProvider::class,
-        ];
-    }
 }